requirements.txt
amqp==5.2.0
APScheduler @ file:///C:/b/abs_eewol9hhy6/croot/apscheduler_1668706376042/work
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.12
django-apscheduler==0.6.2
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.15.1
idna==3.7
kombu==5.3.7
pillow==10.3.0
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
requests==2.31.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal @ file:///C:/b/abs_961jok6k6w/croot/tzlocal_1668627423616/work
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
